Create separate file for init setup
