+++ main.go
@@ -7,10 +7,8 @@ import (
 	"time"
 
 	"github.com/Chingu-cohorts/ChinguDevelopersNetwork/controllers"
-	"github.com/Chingu-cohorts/ChinguDevelopersNetwork/models"
 	"github.com/Chingu-cohorts/ChinguDevelopersNetwork/utils"
 	"github.com/julienschmidt/httprouter"
-	colorable "github.com/mattn/go-colorable"
 	"github.com/phyber/negroni-gzip/gzip"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/rs/cors"
@@ -18,48 +16,6 @@ import (
 	"github.com/urfave/negroni"
 	negroniprometheus "github.com/zbindenren/negroni-prometheus"
 )
-
-func init() {
-	// To use cool colors in windows
-	logrus.SetFormatter(&logrus.TextFormatter{ForceColors: true})
-	logrus.SetOutput(colorable.NewColorableStdout())
-
-	// When initializing the application, we must run the migrations
-	db := utils.InitDB()
-	defer db.Close()
-
-	db.AutoMigrate(&models.Cohort{}, &models.User{}, &models.Project{}, &models.Aptitude{}, &models.Post{}, &models.Comment{})
-
-	// Load cohorts from file
-	cohorts, err := utils.LoadCohortSeed("cohorts.json")
-	if err != nil {
-		logrus.Panic("Could not read the cohorts file")
-	}
-
-	// Iterate over cohorts to save them
-	for _, cohort := range cohorts.Cohorts {
-		db.Create(&cohort)
-
-		logrus.WithFields(logrus.Fields{
-			"cohort": &cohort,
-		}).Info("Saving cohort")
-	}
-
-	// Load aptitudes from file
-	aptitudes, err := utils.LoadAptitudeSeed("aptitudes.json")
-	if err != nil {
-		logrus.Panic("Could not read the aptitudes file")
-	}
-
-	// Iterate over aptitudes to save them
-	for _, aptitude := range aptitudes.Aptitudes {
-		db.Create(&aptitude)
-
-		logrus.WithFields(logrus.Fields{
-			"aptitude": &aptitude,
-		}).Info("Saving aptitude")
-	}
-}
 
 func main() {
 	// Load the config file
